app: Move AppState into its own file and make it global
There are plans to make it a lot bigger. Doing the move now makes it
easier to maintain long running branches that add to it.

Since the store is global, we make the state global too so that it is
easy to use the AppState where we now use the store.  Now you know
where this is going.
